Question 1

+++ Challenge Company 1: Microsoft/Evaluate Reverse Polish Notation.cpp
@@ -28,6 +28,6 @@ public:
                 st.push((long)stoi(it));
             }
         }
-        return (int)st.top();
+        return st.top();
     }
 };
